vue redirect flow and popup flow examples merged
