--- examples/vue-app-redirect-flow/babel.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  presets: ["@vue/cli-plugin-babel/preset"]
-};
