--- examples/vue-app-redirect-flow/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "vue-app-redirect-flow",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
-  },
-  "dependencies": {
-    "@toruslabs/torus-direct-web-sdk": "^4.14.1",
-    "core-js": "^3.8.2",
-    "vue": "^2.6.12"
-  },
-  "devDependencies": {
-    "@types/vue-router": "^2.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.28.2",
-    "@typescript-eslint/parser": "^4.28.2",
-    "@vue/cli-plugin-babel": "^4.5.13",
-    "@vue/cli-plugin-eslint": "^4.5.13",
-    "@vue/cli-plugin-typescript": "~4.5.13",
-    "@vue/cli-service": "^4.5.13",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.26.0",
-    "eslint-plugin-vue": "^7.9.0",
-    "source-map-loader": "^1.1.3",
-    "typescript": "~4.3.5",
-    "vue-template-compiler": "^2.6.12"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ]
-}
