--- examples/vue-app-redirect-flow/src/main.ts
@@ -1,10 +1,0 @@
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount("#app");
