--- examples/vue-app-redirect-flow/vue.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  devServer: {
-    host: "localhost",
-    port: 3000, // CHANGE YOUR PORT HERE!
-    hotOnly: false
-  }
-};
