--- examples/vue-app/.eslintrc.json
@@ -1,16 +1,0 @@
-{
-    "extends": ["plugin:vue/base", "eslint:recommended"],
-    "plugins": ["vue","@typescript-eslint"],
-    "parser": "vue-eslint-parser",
-    "parserOptions": {
-      "parser": "@typescript-eslint/parser",
-      "ecmaVersion": 9,
-      "sourceType": "module"
-    },
-    "env": {
-      "es6": true,
-      "browser": true,
-      "node": true
-    },
-    "root": true
-  }
