+++ examples/vue-app/.prettierrc.yaml
@@ -1,0 +1,5 @@
+# .prettierrc or .prettierrc.yaml
+printWidth: 150
+singleQuote: false
+semi: true
+htmlWhitespaceSensitivity: ignore
