+++ examples/vue-app/package.json
@@ -10,7 +10,8 @@
   "dependencies": {
     "@toruslabs/torus-direct-web-sdk": "^4.14.1",
     "core-js": "^3.8.2",
-    "vue": "^2.6.12"
+    "vue": "^2.6.12",
+    "vue-router": "^3.5.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^4.5.13",
@@ -19,9 +20,11 @@
     "@typescript-eslint/eslint-plugin": "^4.28.2",
     "@typescript-eslint/parser": "^4.28.2",
     "@vue/cli-plugin-typescript": "~4.5.13",
+    "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^7.0.0",
     "typescript": "~4.3.5",
     "babel-eslint": "^10.1.0",
+    "prettier": "^2.3.2",
     "eslint": "^7.26.0",
     "eslint-plugin-vue": "^7.9.0",
     "source-map-loader": "^1.1.3",
