+++ examples/vue-app/src/router/vue-shim.d.ts
